backend: Move `ApartmentHitasAddressSerializer` to utils
