+++ backend/hitas/views/utils/__init__.py
@@ -17,5 +17,10 @@ from hitas.views.utils.filters import (
     HitasUUIDFilter,
 )
 from hitas.views.utils.paginator import HitasPagination
-from hitas.views.utils.serializers import AddressSerializer, HitasAddressSerializer, HitasModelSerializer
+from hitas.views.utils.serializers import (
+    AddressSerializer,
+    ApartmentHitasAddressSerializer,
+    HitasAddressSerializer,
+    HitasModelSerializer,
+)
 from hitas.views.utils.viewsets import HitasModelMixin, HitasModelViewSet
